requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
open3d>=0.13.0
python-docx>=0.8.11
openpyxl>=3.0.7
scikit-learn>=0.24.0 